refactor Assessment code
